Move downloaded file into tmp directory
This commit will mount a tmp directory and move the downloaded file into
it.  This will allow the file to be accessed by the worker service.

+++ hydra/app/jobs/validate_job.rb
@@ -4,16 +4,15 @@ class ValidateJob < ApplicationJob
   queue_as :default
 
   def perform(path:, file_name:, mail_to:)
-    results = ValidationService.new(path: path).validate
-    ## clean up file after processing
-    if File.exist?(path)
-      File.delete(path)
-    end
-    # send the email
-    email_depositor(mail_to:, file_name:, content: results)
+    content = ValidationService.new(path:).validate
+
+    email_depositor(mail_to:, file_name:, content:, path:)
   end
 
-  def email_depositor(mail_to:, file_name:, content:)
+  def email_depositor(mail_to:, file_name:, content:, path:)
+debugger
     ValidationMailer.email_validation(mail_to:, file_name:, content:).deliver_now
+
+    File.delete(path) if File.exist?(path)
   end
 end
